Attempt using libftdi

+++ source/cc3dsfs.cpp
@@ -242,7 +242,7 @@ static void soundCall(AudioData *audio_data, CaptureData* capture_data) {
 				if(data_buffer != NULL) {
 					loaded_samples = audio.samples.size();
 					if((data_buffer->read > get_video_in_size(capture_data)) && (loaded_samples < MAX_MAX_AUDIO_LATENCY)) {
-						int n_samples = get_audio_n_samples(capture_data, data_buffer->read);
+						uint64_t n_samples = get_audio_n_samples(capture_data, data_buffer->read);
 						double out_time = data_buffer->time_in_buf;
 						bool conversion_success = convertAudioToOutput(out_buf[audio_buf_counter], n_samples, endianness, data_buffer, &capture_data->status);
 						if(!conversion_success)
